requirements.txt
pyspark==3.5.2
delta-spark==3.2.0
